requirements.txt
tqdm==4.46.*
beautifulsoup4==4.11.*
requests==2.28.*
pyaml==21.10.*
python-dateutil==2.8.*